Apply clippy's machine-applicable suggestions across the workspace
cargo clippy --fix, over every crate and target: needless borrows and
returns, redundant clones and conversions, iterator shortcuts (contains,
is_some_and, any, first, sort_by_key), collapsed ifs, derived impls in
place of hand-written ones, and single-char patterns. These fixes are
mechanical and leave behaviour unchanged. The warnings that need a
decision follow in their own commits.

+++ crates/fresh-editor/tests/e2e/scrolling.rs
@@ -2114,7 +2114,7 @@ fn test_cursor_visibility_at_line_end_no_wrap() {
     config.editor.line_wrap = false;
     let mut harness = EditorTestHarness::with_config(80, 24, config).unwrap();
 
-    let gutter_width = harness.editor().active_state().margins.left_total_width() as usize;
+    let gutter_width = harness.editor().active_state().margins.left_total_width();
     let visible_width = 80 - gutter_width;
 
     // Create a long line that extends well beyond visible width
